+++ x/accounts/defaults/asset/default.go
@@ -7,6 +7,7 @@ import (
 	v1 "cosmossdk.io/x/accounts/defaults/asset/v1"
 )
 
+// DefaultTransfer same with sdk Transfer logic
 func DefaultTransfer(aa v1.AssetAccountI) func(ctx context.Context, from, to []byte, amount math.Int) ([][]byte, error) {
 	return func(ctx context.Context, from, to []byte, amount math.Int) ([][]byte, error) {
 		err := aa.SubUnlockedCoins(ctx, from, amount)
@@ -23,9 +24,17 @@ func DefaultTransfer(aa v1.AssetAccountI) func(ctx context.Context, from, to []b
 	}
 }
 
+// DefaultMint same with sdk Mint logic
 func DefaultMint(aa v1.AssetAccountI) func(ctx context.Context, to []byte, amount math.Int) ([][]byte, error) {
 	return func(ctx context.Context, to []byte, amount math.Int) ([][]byte, error) {
 		err := aa.AddCoins(ctx, to, amount)
+		if err != nil {
+			return nil, err
+		}
+
+		supply := aa.GetSupply(ctx)
+		supply = supply.Add(amount)
+		err = aa.SetSupply(ctx, supply)
 		if err != nil {
 			return nil, err
 		}
@@ -34,6 +43,7 @@ func DefaultMint(aa v1.AssetAccountI) func(ctx context.Context, to []byte, amoun
 	}
 }
 
+// DefaultBurn same with sdk Burn logic
 func DefaultBurn(aa v1.AssetAccountI) func(ctx context.Context, from []byte, amount math.Int) ([][]byte, error) {
 	return func(ctx context.Context, from []byte, amount math.Int) ([][]byte, error) {
 		err := aa.SubUnlockedCoins(ctx, from, amount)
@@ -41,35 +51,17 @@ func DefaultBurn(aa v1.AssetAccountI) func(ctx context.Context, from []byte, amo
 			return nil, err
 		}
 
+		supply := aa.GetSupply(ctx)
+		supply, err = supply.SafeSub(amount)
+		if err != nil {
+			return nil, err
+		}
+
+		err = aa.SetSupply(ctx, supply)
+		if err != nil {
+			return nil, err
+		}
+
 		return [][]byte{from}, nil
-	}
-}
-
-func CustomTransfer(aa v1.AssetAccountI) func(ctx context.Context, from, to []byte, amount math.Int) error {
-	return func(ctx context.Context, from, to []byte, amount math.Int) error {
-		err := aa.SubUnlockedCoins(ctx, from, amount)
-		if err != nil {
-			return err
-		}
-
-		fee := math.LegacyNewDecWithPrec(10, 2) // 10%
-		feeAmount := math.LegacyNewDecFromInt(amount).Mul(fee).TruncateInt()
-		transferAmount := amount.Sub(feeAmount)
-		owner, err := aa.GetOwner(ctx)
-		if err != nil {
-			return err
-		}
-
-		err = aa.AddCoins(ctx, to, transferAmount)
-		if err != nil {
-			return err
-		}
-
-		err = aa.AddCoins(ctx, owner, feeAmount)
-		if err != nil {
-			return err
-		}
-
-		return nil
 	}
 }
